backend/requirements/base.txt
Django==5.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.*
django-environ==0.11.*
django-cors-headers==4.*
psycopg[binary]==3.*
redis==5.*
celery[redis]==5.*
django-celery-beat==2.*
gunicorn==22.*
httpx==0.28.*
django-allauth[socialaccount]==65.*
dj-rest-auth[with_social]==7.*
